# public/openapi.yaml
openapi: 3.0.3
info:
  title: 同频 B2 API
  version: 1.3.0
  description: 同频 B2 情绪急救与数字陪伴前后端接口。支持 Swagger UI Try it out；可在 Servers 中切换本地或远程环境。注意：本 API 不替代专业医疗或心理服务，AI 输出必须人工确认。
servers:
- url: http://localhost:8080/api/v1
  description: 本地开发环境
- url: https://{remoteHost}/api/v1
  description: 远程环境（修改 remoteHost 为实际域名）
  variables:
    remoteHost:
      default: api.tongpin.example.com
      description: 远程 API 域名，不带 https:// 和路径
...
paths:
...
  /resources:
    get:
      tags:
      - 公共资源
      summary: 获取现实求助渠道
      operationId: listResources
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /manager/overview:
    get:
      tags:
      - 管理端
      summary: 获取辖区总览
      operationId: getManagerOverview
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/users:
    get:
      tags:
      - 管理端
      summary: 获取辖区用户摘要
      operationId: listManagerUsers
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/volunteers:
    get:
      tags:
      - 管理端
      summary: 获取志愿者与负荷
      operationId: listManagerVolunteers
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/schedules:
    get:
      tags:
      - 管理端
      summary: 获取排班
      operationId: listManagerSchedules
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/schedules/{schedule_id}:
    patch:
      tags:
      - 管理端
      summary: 调整排班
      operationId: updateManagerSchedule
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
        description: schedule_id 标识
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              start_at: '2026-09-09T08:00:00+08:00'
              end_at: '2026-09-09T12:00:00+08:00'
              role_in_shift: primary
              confirmed: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/transfer-requests:
    get:
      tags:
      - 管理端
      summary: 获取交接队列
      operationId: listTransferRequests
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferRequest'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/transfer-requests/{id}/confirm:
    post:
      tags:
      - 管理端
      summary: 确认普通交接
      operationId: confirmTransferRequest
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: id 标识
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              receiver_volunteer_id: volunteer_02H8YQ
              capacity_verified: true
              note: 已核对接班人剩余名额和资格。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferRequest'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/professional-requests:
    get:
      tags:
      - 管理端
      summary: 获取专业接管请求
      operationId: listProfessionalRequests
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/professional-requests/{id}/confirm:
    post:
      tags:
      - 管理端
      summary: 确认专业接管
      operationId: confirmProfessionalRequest
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: id 标识
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              reviewed_user_situation: true
              takeover_reason: 用户明确表示当前不安全。
              next_action: coordinate_external_referral
              external_referral_status: pending
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/support-reviews:
    post:
      tags:
      - 管理端
      summary: 记录支持分层复核
      operationId: createSupportReview
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              case_id: case_01H8YQ
              support_need_level: sustained_support
              evidence_summary: 本人反馈持续低落并影响睡眠。
              valid_until: '2026-09-16T00:00:00+08:00'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 输入登录后获取的 access token。匿名用户先调用 /auth/user-session。
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: 写接口幂等键。重复提交同一请求时应返回同一结果。
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
...
    TransferRequest:
      type: object
      properties:
        id:
          type: string
        case_id:
          type: string
        type:
          type: string
        from_volunteer_id:
          type: string
        to_volunteer_id:
          type:
          - string
          - 'null'
        status:
          type: string
        receiver_confirmed_at:
          type:
          - string
          - 'null'
          format: date-time
    Resource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        region:
          type: string
        phone:
          type: string
        available_time:
          type: string
        official_source:
          type: string
        verified_at:
          type: string
          format: date-time
        status:
          type: string
...
    GenericObject:
      type: object
      additionalProperties: true